converting new stream_svc into a contact (#1228)
* converting new stream_svc into a contact

* converting new stream_svc into a contact

* ensuring all beacons include a tag, for udp

* ensuring all beacons include a tag, for udp
