+++ app/contacts/contact_http.py
@@ -9,7 +9,7 @@ class Http(BaseWorld):
 
     def __init__(self, services):
         self.name = 'http'
-        self.description = 'Communication occurs through REST API endpoints'
+        self.description = 'Accept beacons through a REST API endpoint'
         self.app_svc = services.get('app_svc')
         self.contact_svc = services.get('contact_svc')
